moved the additional functions to functions.py
